requirements.txt
pymusic-dl
setproctitle
pydub
pyaudio
dataset
pyyaml